Move codec to a separate directory
