+++ oak/server/rust/oak_runtime/src/node/grpc/mod.rs
@@ -14,4 +14,5 @@
 // limitations under the License.
 //
 
+mod codec;
 pub mod server;
